move ingestpingdata from compass
